tg/requirements.txt
aiogram>=3.4,<4.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<0.28
